refactor: Avoid multiple API calls for location types
This refactors how location types are fetched from the API. Now that
it's possible to make a single request with multiple location types, we
don't need to make multiple API requests.

I've also removed the functionality where we sort the result and we can
now rely on the API returning the results in a suitably sorted manner.
This does mean one of the tests has changed to not have the output in a
sorted way, but I think this is a better as it makes it explicit that
the frontend is not performing any kind of sorting on the results from
the API.

Closes #1028