chore: Warn when the auth source cannot be determined
This follows on from
https://github.com/ministryofjustice/hmpps-book-secure-move-frontend/pull/1790
to add more information around missing location errors. We hypothesis
that perhaps the reason the user has no locations is because the auth
source was unable to be determined.

By warning in this situation we're hoping to be able to figure out in
more detail what's going on.