refactor: Extract supplier user location fetching
This pulls out the method for fetching and updating the locations for a
supplier user. The overall behaviour isn't different, but it will make it
easier to move this method into the authentication process in the future.
This is something we want to do as we think that supplier users are
losing their locations when their authentication session needs
regenerating.