fix: Populate supplier user locations on session regeneration
This ensures that when regenerating a session, if the user is a
supplier, the locations are fetched. Previously this was happening in
the locations controller.

Because of this, it was only working correctly because users end up on
the locations controller after they first log in, but if a user needs to
be re-authenticated in the background, the locations would be lost.