refactor: Set flag on move to determine whether it is editable
Currently there are a number of scenarios that allow editing of a move,
including its state and whether a user has permission to edit.

This change extracts that logic to a separate helper and moves it
to the service which has access to the request, and therefore the
user and their permissions.