refactor: Set whether a move can be cancelled within the service
This follows a similar pattern to determining whether a move can be
cancelled to the service where the move is returned.

This logic is based on both a user's permissions and the state of the
move.

Moving this logic to this location means that whenever a move is
accessed by itself it will have a property saying whether it can be
cancelled.