build: Add latest frameworks package
This includes a new type of NOMIS mapping for assessments, a new
religious and cultural question and some further grammar changes.

We don't want to start using this version until the suppliers are
confident using it themselves, so we'll need to set the
`FRAMEWORKS_VERSION` environment variable in production to the previous
version.