fix: Prevent metrics library from erroring without correct type
When some API requests fail we are seeing a
`Value is not a valid number: NaN` error which suggests one of the
values we're sending to this library isn't valid.

This fix checks for a valid end timing, whether a `response.end` or
`request.error` value, and won't call the metrics library if this
value doesn't exist.

Alongside that, some extra context has been added to the errors to
help us determine which values are incorrect.

It also fixes an issue with the correct status code not being sent
to the errors middleware meaning the code displayed on the error page
when in development mode wasn't correct.

fixes BOOK-A-SECURE-MOVE-FRONTEND-5E