refactor: Reduce duplication date string rendering
This pulls out some common code into a function and calls it in multiple
places to avoid duplication. There is a slightly change to the output
for one of the calls, which is that there is an extra empty line, but I
don't think that's enough to avoid this refactor.