fix: Fix missing background images
This relates to a recent version change to `css-loader` which caused
the images to no longer be created with the expected filenames or
in the correct directory.

The latest version of css-loader now uses Webpack 5's asset modules,
whilst our configuration was using the legacy Webpack 4 method of
`file-loader`.

This fix updates the configuration to use the new method whilst keeping
the existing directory and filename structure.