fix: Ensure cache is cleared on errors during deserialization
Currently devour, the JSON:API client will build a cached of
deserialized resources. We have added a transformation layer to help
with certain things the frontend requires for these resources.

However, if an error occurs, the cache remains in tact which will
mean that in some cases, on a refresh of the page the app will use
the cached version.

This can lead to some unexpected renders so this fix calls the `clear()`
method on the cache to ensure on the next request it will begin again
from the start.