fix: Ensure move details aren't shown on form wizard unless present
This fixes an issue where a move summary might not be available but
the move details heading will still appear.

This is currently the case for the creation of new allocations.

I did investigate seeing if we could pull in the allocation details
during this journey instead but it would've invovled a little more
work at this stage so have decided to revert back to the previous
implementation for now where there is no side panel.