fix: Support prison recall moves that also have a destination
This updates the presenters where a destination is shown as well
as the move details edit step.

Previously prison recall moves were created without a destination (or
`to_location`).

So most of the logic on the frontend was driven just by `move_type`.

However, now suppliers are updating the API with the actual end
destination we need to adapt the logic on the frontend for both
rendering and editing to cater for scenarios where `prison_recall`
moves have a `to_location`.

fixes BOOK-A-SECURE-MOVE-FRONTEND-55