feat: Reduce PNC validation to Sentry warnings only
This change removes fields validated with the new Police National
Computer (PNC) validator from the form wizard so that the user doesn't
see this error.

It then adds each instance to Sentry as a warning.

This is because we are not fully certain of the impact of introducing
this validation yet. There are some establishments that are using this
field for non-PNC reference numbers and we want to get an understanding
of scale first before enabling it.

These warnings will allow us to see the location types that are
causing this validation to trigger and allow us to decide an action
to take.