chore: Update "Change location" copy to match DPS
This is a minor tweak to the copy on the "Change location" action in
the organisation switcher bar.

Digital Prison Services uses a tweaked version and to bring us more
inline this is being updated. This will help make users feel like they
are accessing services that are all part of the same space.