refactor: Use tabs template for old move page
This updates the tabs on the old (current) move page to use the new tabs
template. This requires a bit of refactoring to support the variable
expected by the tabs, and I've decided to not refactor it completely to
middleware as I'm not expecting these pages to be around for much
longer.