feat: Redirect person root to personal-details
This introduces a redirect from the person root URL to the personal
details tag. This matches the behaviour of the new move page and sets us
up for this page having multiple tabs.

To support this, I've also changed how the breadcrumbs are set, so the
personal details page now appears as one level down from the "person
page" and this will facilitate having multiple tabs in the future.