chore: Don't only record performance data in production
Having the data in the other environments allows us to be aware of any
potential problems if, for example, a new release in staging suddenly
causes requests to take longer to respond.

This shouldn't increase the number of traces we record and therefore
shouldn't affect rate limits. It also reduces differences in the app in
different environments which makes debugging easier.