fix: Prevent journeys without a vehicle throwing an error
When this code was first implemented it was using destructuring for
vehicle fallback.

But in the cases where a vehicle doesn't exist in the API it can
return a value of `null` which prevents the fallback object from being
set as the value.

This meant that the rest of the presenter would error trying to access
properties of a `null` value.

This change ensures the presenter doesn't error when missing a vehicle
or when the vehicle value is set to `null`.

It also adds fallback text for when there is no vehicle.

We don't need to do the same for `from_location` and `to_location`
as the API will not accept a journey without these values.