chore: Create temporary error routes to debug Sentry
This creates 2 new routes to intentionally trigger an error when
accessing a property of `null`.

We have seen this error in production but it was not captured in
Sentry. In order to rule out anything to do with limits or new errors
not being captured this change adds a way to throw a similar error.

This change should be reverted once this problem has been resolved.