chore: Make it possible to debug Sentry
This adds an optional environment variable (`SENTRY_DEBUG`) which can
be set to `true` and will turn on Sentry debugging. This allows us to
investigate why issues might not be appearing in Sentry.

It's not good practice to set this in production for long periods of
time, but it doesn't pose any risk of leaking information.

See
https://docs.sentry.io/platforms/javascript/configuration/options/#debug
for more information on the option.