chore: Don't collect locations in Sentry issues
This partially reverts a9f1f5384ae3fdee693cadf5684051b5993a6888 to stop
sending location information to Sentry. We've discovered that Sentry
errors triggered from members of our team have so many locations that
the payload is too big to be accepted by Sentry, so we get 413 errors
and no issue appearing in Sentry.

We only added this extra information to debug a separate issue which has
since been fixed, so I think we can remove it. If we ever need it back
again we should think about only exporting the IDs of the locations
rather than the whole object.