feat: Add banner to opt-in to preview new move design
This creates the basic routes to handle opting in and out of the preview
design for the move page as well as the banner at the top of the
existing move page design.

The routes set a cookie with the user's ID to cater for multiple users
using the same machine as well as their answer.

This will be used to either redirect the user if they have opted in
or continue to show them the existing design if they opted out.