refactor: Replace individual move change links with single link
This reduces the amount of actions currently being shown in the move
summary sidebar.

There are a number of other things that can be edited for a move and
also just viewed. So in order to help users find all of those details
the individual change links are being replaced with one in the side
panel that will take them to the screen to view all the details for
this move.