fix: Ensure only select routes are redirect after opting in to preview
This fixes a bug where the redirect was occuring for all routes once
a user had opted in.

Some routes still need to remain on the existing paths and shouldn't
be redirected. This change passes the expected routes in to the
middleware method so that they are checked first.

The `router.use` method doesn't accept an array as the first path which
is why this functionality wasn't behaving as expected previously.

It also uses an object so that it can cater for any routes that have
been renamed in the new version.