refactor: Don't pass whole person object to views via person summary
This refactors the code that passes the person summary to the view
layer.

Previously it would pass the entire `person` object so that the view
could do logic to show the image and profile link if it exists.

This was a bit unnecessary so this change removes the whole object
and only passes the info it needs. This also allows us to only show
particular elements, like the profile link, in certain circumstances.