refactor: Don't pass move with move summary helper
This move summary helper also returns the whole move object. This
feels overkill as its already passed in via the `get-locals` method
and means sending the whole move object multiple times.

It would also make it harder to debug.

This function is renamed as part of this to be more descriptive.