refactor: Use shared middleware to load move summary panel
This refactors the places where the move summary is loaded via
middlware and assigned to the view locals.

This is done in various ways across the code base. This change
standardises this to a single middleware that already exists which
will allow more elements to be added and shared across all instances.