fix: Fix missing allocation cancelled description
The `sending_establishment_failed_to_fill_allocation` cancellation
reason was actually listed as `unfilled` in the locale file, which meant
this wasn't being rendered on the frontend when it was the case.

I've also re-order the cancellation reasons to match the ordering in the
API so it's slightly easier to match up.