fix: Remove youth_risk_assessment:print permission
This isn't an action that can be performed on youth risk assessments, so
there's no reason to have a permission for it.

By removing this permission it should avoid the "Print Youth Risk
Assessment" links from appearing on the frontend as they only render if
the user has the permission. This fixes a bug where pressing that link
would take the user to a 404 page.