feat: Change target of "Save and return to overview" button
This changes the link target of the "Save and return to overview" button
of a framework step to send the user to the framework overview, rather
than the framework section.

We'd like to change this to be consisten on the naming of "overview" so
users better understand where they are in the journey.