feat: Combine fullname and reference in move link
This makes the move links clearer where they are taking the user to, by
combining the person full name and move reference on to one line. At the
moment the link is just the person's name which can be confusing as
you'd expect that to take you to the person instead of the move.