feat: Improve "Save and continue" navigation
This changes the button to read "Save and return to overview" when on
the last step of the last section of a framework. It also hides the
"Save and return to overview" button which is visible on all other
pages.

The idea here is to improve the navigation for users so they don't have
to rely on the section "overview" pages as much.