feat: Always show filter boxes in a collection view
The current behaviour hides the filter boxes when there are no results,
but this can be confusing for users as there's no indication that there
are no results elsewhere on the page. By keeping the filter boxes shown,
the page looks the same regardless of how many results there are.