fix: Ensure text is escaped before being rendered
This prevents potential XSS security vulnerabilities where users can
enter HTML in to a field and then that HTML is rendered on the frontend.
This is particularly dangerous if a user inputs a `<script>` tag in to
the field.

The vulnerability is limited because you have to be authenticated to
access the page.