refactor: Allow breadcrumbs block to be customised
By moving the breadcrumbs in to a block this ensures we can completely
replace it if we need to customise this behaviour. There is another way
of customising this which is to remove the breadcrumbs, however this
poses a problem as the breadcrumbs are used for reasons other than just
visual links (the title of the page is determined using the last
breadcrumb).