fix: Replace breadcrumbs in new move design
We're going to replace the breadcrumbs with a link to the dashboard as
per the old design as this has been identified as a blocker to releasing
the new design.

Once the design is launcher, we will look at improving the breadcrumbs
so they work better and we can remove this replacement.