fix: Make it possible to edit personal details in new design
We've spotted an issue with the new move page design where it's no
longer possible to change the person details of a person from the move
page. This behaviour has been lost from the old design, so we're putting
in a quick fix now to allow us to continue with the launch.