feat: Include property change events in move overview
We'd like to include these events as important events that happen during
a move. Since these events are not classified as incidents, the code had
to be changed slightly to support these.

The code is a little messy, as it doesn't actually rely on the important
events field as its returned from the API, but I think refactoring this
is for a different time.