fix: Ensure label is found when adding another
This copies the fix from MoJ Frontend
(https://github.com/ministryofjustice/moj-frontend/pull/186) in to our
own copy of the library, see details on that original PR for what the
issue is and why this fixes it.

Ideally, we wouldn't have customised this component so much that
upstream fixes are applied downstream, but for now at least this fixes
the bug.