fix: Consider assessment status when showing framework responses
Currently we always show a link to amend a framework response if the
value is not set, regardless of the assessment status. This causes a
problem with optional top-level questions which may not have been
responded to by the time the PER is complete. At that point we can't let
the user modify the PER though, as the PER cannot be changed once
complete.