feat: Make sure newly created moves have a PER
Sometimes PERs dont get started by the time someone ends up in custody. For
example sometimes someone is in court but no PER events can be added by the
supplier because the PER has not been created. In order to help alleviate this,
we have decided to create an empty PER object whenever a move is created, so
that events can be added to it before it has been started or completed.

[P4-4174]